+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* **New** - WooCommerce: customer the order received title page when payment failed
+
 ### 1.2.1: September 21st, 2020
 * **New** - Oembed: YouTube videos always with rel=0 modestbranding=1 showinfo=0